+++ src/elements/prto.c
@@ -52,8 +52,8 @@ int update_PRTO(UPDATE_FUNC_ARGS) {
 							np = create_part(-1,x+rx,y+ry,portalp[parts[i].tmp][randomness][nnx].type);
 							if (np<0) continue;
 							parts[np] = portalp[parts[i].tmp][randomness][nnx];
-							parts[np].x = x+rx;
-							parts[np].y = y+ry;
+							parts[np].x = (float)x+rx;
+							parts[np].y = (float)y+ry;
 							portalp[parts[i].tmp][randomness][nnx] = emptyparticle;
 							break;
 						}
